# cloudmall-admin/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: renren-fast 是一個轻量級的 Java 快速開發平台，能快速開發項目並交付【接私活利器】
  version: 1.0.0
  title: 人人快速開發平台
basePath: /renren-fast
schemes:
  - http

#認証
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定義接口資料
paths:
...
  /sys/user/list:
    get:
      tags:
        - 管理員管理
      summary: 管理員列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: username
          description: 使用者名稱
          in: query
          type: string
      responses:
        '200':
          description: 返回管理員列表
          schema:
            $ref: '#/definitions/SysUserEntityList'
...
  /sys/user/info/{userId}:
    get:
      tags:
        - 管理員管理
      summary: 取得管理員資訊
      produces:
        - application/json
      parameters:
        - name: userId
          description: 使用者ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回管理員資訊
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserEntity'
  /sys/user/password:
    post:
      tags:
        - 管理員管理
      summary: 修改密碼
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/PasswordForm'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/save:
    post:
      tags:
        - 管理員管理
      summary: 新增管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/update:
    post:
      tags:
        - 管理員管理
      summary: 修改管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 管理員物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysUserEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/user/delete:
    post:
      tags:
        - 管理員管理
      summary: 删除管理員
      produces:
        - application/json
      parameters:
        - name: body
          description: 使用者ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 選單管理
      summary: 取得選單備註
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 選單ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回選單資訊
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              menu:
                description: 選單資訊
                $ref: '#/definitions/SysMenuEntity'
  /sys/menu/save:
    post:
      tags:
        - 選單管理
      summary: 新增選單
      produces:
        - application/json
      parameters:
        - name: body
          description: 選單物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - 選單管理
      summary: 修改選單
      produces:
        - application/json
      parameters:
        - name: body
          description: 選單物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete/{menuId}:
    post:
      tags:
        - 選單管理
      summary: 删除選單
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 選單ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/config/list:
    get:
      tags:
        - 参數管理
      summary: 参數列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: key
          description: 参數名
          in: query
          type: string
      responses:
        '200':
          description: 返回参數列表
          schema:
            $ref: '#/definitions/SysConfigEntityList'
  /sys/config/info/{id}:
    get:
      tags:
        - 参數管理
      summary: 取得参數資訊
      produces:
        - application/json
      parameters:
        - name: id
          description: 参數ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回参數資訊
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              config:
                description: 返回参數資訊
                $ref: '#/definitions/SysConfigEntity'
  /sys/config/save:
    post:
      tags:
        - 参數管理
      summary: 新增参數
      produces:
        - application/json
      parameters:
        - name: body
          description: 参數物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysConfigEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/update:
    post:
      tags:
        - 参數管理
      summary: 修改参數
      produces:
        - application/json
      parameters:
        - name: body
          description: 参數物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysConfigEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/delete:
    post:
      tags:
        - 参數管理
      summary: 删除参數
      produces:
        - application/json
      parameters:
        - name: body
          description: 参數ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/oss/list:
    get:
      tags:
        - 檔案服務
      summary: 檔案列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
      responses:
        '200':
          description: 返回檔案列表
          schema:
            $ref: '#/definitions/SysOssEntityList'
...
  /sys/oss/saveConfig:
    post:
      tags:
        - 檔案服務
      summary: 保存雲端存儲設定資訊
      produces:
        - application/json
      parameters:
        - name: body
          description: 参數物件
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysCloudStorageEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/oss/upload:
    post:
      tags:
        - 檔案服務
      summary: 上傳檔案
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: file
          description: 檔案
          in: formData
          type: file
          required: true
      responses:
        '200':
          description: 返回檔案列表
          schema:
            $ref: '#/definitions/FileUpload'
  /sys/oss/delete:
    post:
      tags:
        - 檔案服務
      summary: 删除檔案
      produces:
        - application/json
      parameters:
        - name: body
          description: 檔案ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: 狀態碼 0：成功  非0：失敗
          type: integer
          format: int32
        msg:
          description: 失敗原因
          type: string
...
  PasswordForm:
    type: object
    properties:
      password:
        description: 原密碼
        type: string
      newPassword:
        description: 新密碼
        type: string
  SysUserEntity:
    type: object
    properties:
      userId:
        description: 使用者ID
        type: integer
        format: int64
      username:
        description: 使用者名稱
        type: string
      password:
        description: 密碼
        type: string
      email:
        description: 電子郵件
        type: string
      mobile:
        description: 手機號
        type: string
      status:
        description: 狀態  0：禁用   1：正常
        type: integer
        format: int32
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 創建者 id
        type: integer
        format: int64
      createTime:
        description: 創建時間
        type: string
        format: date-time
  SysUserEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserEntity'
  SysUserEntityEdit:
      type: object
      properties:
        userId:
          description: 使用者ID
          type: integer
          format: int64
        username:
          description: 使用者名稱
          type: string
        password:
          description: 密碼
          type: string
        email:
          description: 電子郵件
          type: string
        mobile:
          description: 手機號
          type: string
        status:
          description: 狀態  0：禁用   1：正常
          type: integer
          format: int32
        roleIdList:
          description: 角色ID列表
          type: array
          items:
            type: integer
            format: int32
...
  SysMenuEntity:
    type: object
    properties:
      menuId:
        description: 選單ID
        type: integer
        format: int64
      name:
        description: 選單名稱
        type: string
      parentId:
        description: 父選單ID，一級選單為0
        type: integer
        format: int64
      parentName:
        description: 父選單名稱
        type: string
      url:
        description: 選單URL
        type: string
      perms:
        description: 授權標識
        type: string
      type:
        description: 類型     0：目錄   1：選單   2：按钮
        type: integer
        format: int32
      icon:
        description: 選單圖標
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展開   true：展開   false：不展開
        type: boolean
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 選單ID
        type: integer
        format: int64
      name:
        description: 選單名稱
        type: string
      parentId:
        description: 父選單ID，一級選單為0
        type: integer
        format: int64
      url:
        description: 選單URL
        type: string
      perms:
        description: 授權標識
        type: string
      type:
        description: 類型     0：目錄   1：選單   2：按钮
        type: integer
        format: int32
      icon:
        description: 選單圖標
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...
  SysConfigEntity:
    type: object
    properties:
      id:
        description: 参數ID
        type: integer
        format: int64
      key:
        description: 参數名
        type: string
      value:
        description: 参數值
        type: string
      remark:
        description: 備註
        type: string
  SysConfigEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysConfigEntity'
  SysOssEntity:
    type: object
    properties:
      id:
        description: ID
        type: integer
        format: int64
      url:
        description: URL地址
        type: string
      createTime:
        description: 創建時間
        type: string
        format: date-time
  SysOssEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysOssEntity'
  SysCloudStorageEntity:
    type: object
    properties:
      type:
        description: 類型 1：七牛  2：阿里雲  3：腾訊雲
        type: integer
        format: int32
      qiniuDomain:
        description: 七牛绑定的域名
        type: string
      qiniuPrefix:
        description: 七牛路径前缀
        type: string
      qiniuAccessKey:
        description: 七牛ACCESS_KEY
        type: string
      qiniuSecretKey:
        description: 七牛SECRET_KEY
        type: string
      qiniuBucketName:
        description: 七牛存儲空間名
        type: string
      aliyunDomain:
        description: 阿里雲绑定的域名
        type: string
      aliyunPrefix:
        description: 阿里雲路径前缀
        type: string
      aliyunEndPoint:
        description: 阿里雲EndPoint
        type: string
      aliyunAccessKeyId:
        description: 阿里雲AccessKeyId
        type: string
      aliyunAccessKeySecret:
        description: 阿里雲AccessKeySecret
        type: string
      aliyunBucketName:
        description: 阿里雲BucketName
        type: string
      qcloudDomain:
        description: 腾訊雲绑定的域名
        type: string
      qcloudPrefix:
        description: 腾訊雲路径前缀
        type: string
      qcloudAppId:
        description: 腾訊雲AppId
        type: string
      qcloudSecretId:
        description: 腾訊雲SecretId
        type: string
      qcloudSecretKey:
        description: 腾訊雲SecretKey
        type: string
      qcloudBucketName:
        description: 腾訊雲 BucketName
        type: string
      qcloudRegion:
        description: 腾訊雲 COS 所屬地區
        type: string
  FileUpload:
      type: object
      properties:
        code:
          description: 狀態碼 0：成功  非0：失敗
          type: integer
          format: int32
        url:
          description: 檔案URL地址
          type: string
        msg:
          description: 失敗原因
          type: string
...